@@ -3,11 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-particles-js": "^2.1.0",
-    "react-scripts": "^4.0.3",
-    "react-tilt": "^0.1.4",
+    "particles-bg": "^2.5.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-parallax-tilt": "^1.7.67",
+    "react-scripts": "^5.0.1",
     "tachyons": "^4.12.0"
   },
   "scripts": {
