@@ -1,9 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 // import registerServiceWorker from './registerServiceWorker';
 import 'tachyons';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+
+root.render(<App />)
+
+// If using React version lower than 18:
+// ReactDOM.render(<App />, document.getElementById('root'));
 // registerServiceWorker();
