@@ -1,5 +1,6 @@
 import React from 'react';
-import Tilt from 'react-tilt';
+// import Tilt from 'react-tilt'; 
+import Tilt from 'react-parallax-tilt';
 import brain from './brain.png';
 import './Logo.css';
 
